+++ child/testutil/simapp/simapp.go
@@ -15,7 +15,7 @@ import (
 	tmtypes "github.com/tendermint/tendermint/types"
 	tmdb "github.com/tendermint/tm-db"
 
-	"github.com/cosmos/interchain-security/app"
+	"github.com/cosmos/interchain-security/child/app"
 )
 
 // New creates application instance with in-memory database and disabled logging.
