+++ x/ccv/child/keeper/keeper_test.go
@@ -14,8 +14,8 @@ import (
 	channeltypes "github.com/cosmos/ibc-go/modules/core/04-channel/types"
 	commitmenttypes "github.com/cosmos/ibc-go/modules/core/23-commitment/types"
 	ibctmtypes "github.com/cosmos/ibc-go/modules/light-clients/07-tendermint/types"
-	"github.com/cosmos/interchain-security/app"
-	"github.com/cosmos/interchain-security/testutil/simapp"
+	"github.com/cosmos/interchain-security/child/app"
+	"github.com/cosmos/interchain-security/child/testutil/simapp"
 	childtypes "github.com/cosmos/interchain-security/x/ccv/child/types"
 	parenttypes "github.com/cosmos/interchain-security/x/ccv/parent/types"
 	"github.com/cosmos/interchain-security/x/ccv/types"
