+++ x/ccv/child/module_test.go
@@ -11,8 +11,8 @@ import (
 	host "github.com/cosmos/ibc-go/modules/core/24-host"
 	ibctmtypes "github.com/cosmos/ibc-go/modules/light-clients/07-tendermint/types"
 	ibctesting "github.com/cosmos/ibc-go/testing"
-	"github.com/cosmos/interchain-security/app"
-	"github.com/cosmos/interchain-security/testutil/simapp"
+	"github.com/cosmos/interchain-security/child/app"
+	"github.com/cosmos/interchain-security/child/testutil/simapp"
 	"github.com/cosmos/interchain-security/x/ccv/child"
 	childtypes "github.com/cosmos/interchain-security/x/ccv/child/types"
 	parenttypes "github.com/cosmos/interchain-security/x/ccv/parent/types"
